refactor(napi): share panic error conversion
