+++ crates/napi/src/bindgen_runtime/mod.rs
@@ -9,7 +9,7 @@ pub use js_values::*;
 pub use module_register::*;
 
 use super::sys;
-use crate::{JsError, Result, Status};
+use crate::{Error, JsError, Result, Status};
 
 #[cfg(feature = "tokio_rt")]
 pub mod async_iterator;
@@ -28,6 +28,20 @@ pub trait ObjectFinalize: Sized {
   fn finalize(self, env: Env) -> Result<()> {
     Ok(())
   }
+}
+
+#[doc(hidden)]
+pub fn panic_to_error(e: Box<dyn std::any::Any + Send>) -> Error {
+  let message = {
+    if let Some(string) = e.downcast_ref::<String>() {
+      string.clone()
+    } else if let Some(string) = e.downcast_ref::<&str>() {
+      string.to_string()
+    } else {
+      format!("panic from Rust code: {:?}", e)
+    }
+  };
+  Error::new(Status::GenericFailure, message)
 }
 
 /// # Safety
